Refactor | add indexoutofbounds package and test method rename.
