+++ core-java-modules/core-java-exceptions-3/src/main/java/com/baeldung/indexoutofbounds/CopyListUsingAddAllMethodDemo.java
@@ -1,10 +1,9 @@
-package com.baeldung;
+package com.baeldung.indexoutofbounds;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class CopyListUsingAddAllMethodDemo {
-
     static List<Integer> copyList(List<Integer> source) {
         List<Integer> destination =  new ArrayList<>();
