+++ core-java-modules/core-java-exceptions-3/src/main/java/com/baeldung/indexoutofbounds/CopyListUsingCollectionsCopyMethodDemo.java
@@ -1,10 +1,9 @@
-package com.baeldung;
+package com.baeldung.indexoutofbounds;
 
 import java.util.Collections;
 import java.util.List;
 
 public class CopyListUsingCollectionsCopyMethodDemo {
-
     static void copyList(List<Integer> source, List<Integer> destination) {
         Collections.copy(destination, source);
     }
