+++ core-java-modules/core-java-exceptions-3/src/main/java/com/baeldung/indexoutofbounds/CopyListUsingConstructorDemo.java
@@ -1,12 +1,10 @@
-package com.baeldung;
+package com.baeldung.indexoutofbounds;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class CopyListUsingConstructorDemo {
-
     static List<Integer> copyList(List<Integer> source) {
-
         return new ArrayList<>(source);
     }
 }
