+++ core-java-modules/core-java-exceptions-3/src/main/java/com/baeldung/indexoutofbounds/CopyListUsingJava8StreamDemo.java
@@ -1,10 +1,9 @@
-package com.baeldung;
+package com.baeldung.indexoutofbounds;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 public class CopyListUsingJava8StreamDemo {
-
     static List<Integer> copyList(List<Integer> source) {
         return source
                 .stream()
