+++ core-java-modules/core-java-exceptions-3/src/test/java/com/baeldung/indexoutofbounds/CopyListUsingCollectionsCopyMethodDemoUnitTest.java
@@ -1,32 +1,31 @@
-package com.baeldung;
+package com.baeldung.indexoutofbounds;
 
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
 import java.util.List;
 
-import static com.baeldung.CopyListUsingCollectionsCopyMethodDemo.copyList;
 import static org.junit.jupiter.api.Assertions.*;
 
 class CopyListUsingCollectionsCopyMethodDemoUnitTest {
 
     @Test
-    void shouldCopySourceListElementsInDestinationList() {
+    void whenCopyListUsingCollectionsCopy_thenOverrideAllDestinationListValues() {
         List<Integer> source = Arrays.asList(11, 22, 33);
         List<Integer> destination = Arrays.asList(1, 2, 3);
 
-        copyList(source, destination);
+        CopyListUsingCollectionsCopyMethodDemo.copyList(source, destination);
 
         assertEquals(source, destination);
     }
 
     @Test
-    void shouldOverrideSourceListElementsToDestinationList() {
+    void whenCopyListUsingCollectionsCopy_thenOverrideInitialDestinationValuesAndOthersShouldBeUnchanged(){
         List<Integer> source = Arrays.asList(11, 22, 33);
         List<Integer> destination = Arrays.asList(1, 2, 3, 4, 5);
         List<Integer> expectedList = Arrays.asList(11, 22, 33, 4, 5);
 
-        copyList(source, destination);
+        CopyListUsingCollectionsCopyMethodDemo.copyList(source, destination);
 
         assertEquals(expectedList, destination);
     }
