+++ core-java-modules/core-java-exceptions-3/src/test/java/com/baeldung/indexoutofbounds/CopyListUsingConstructorDemoUnitTest.java
@@ -1,19 +1,18 @@
-package com.baeldung;
+package com.baeldung.indexoutofbounds;
 
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
 import java.util.List;
 
-import static com.baeldung.CopyListUsingConstructorDemo.copyList;
 import static org.junit.jupiter.api.Assertions.*;
 
 class CopyListUsingConstructorDemoUnitTest {
 
     @Test
-    void shouldMakeACopyOfSourceListUsingArrayListConstructor() {
+    void whenCopyListUsingConstructor_thenMakeACopyOfList() {
         List<Integer> source = Arrays.asList(11, 22, 33);
 
-        assertEquals(source, copyList(source));
+        assertEquals(source, CopyListUsingConstructorDemo.copyList(source));
     }
 }
